pyproject.toml
# Python project configuration for AIRAS prompt-tuning experiments

[project]
name = "airas-prompt-tuning"
version = "0.1.0"
description = "Prompt-tuning experiments for improved Chain-of-Thought reasoning"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    
    # ML frameworks
    "torch>=2.1.0",
    "datasets>=2.14.0",
    
    # LLM APIs
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    
    # Data processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # Utilities
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
